+++ config/routes.rb
@@ -40,6 +40,8 @@ Upright::Engine.routes.draw do
 
     resources :artifacts, only: :show, as: :site_artifacts
 
+    match "traces/:signed_id", to: "traces#show", as: :site_trace, via: [ :get, :options ]
+
     scope :framed do
       resource :jobs, only: :show
     end